Entity定義をUserとProfileでファイルに分けた
