+++ pkg/domain/entity/profile.go
@@ -1,0 +1,16 @@
+package entity
+
+import "time"
+
+type Profile struct {
+	Name      string
+	Thumbnail string
+	Bio       string
+	Gender    int
+	Phone     string
+	Place     string
+	Birth     string
+	CreatedAt *time.Time
+	UpdatedAt *time.Time
+	DeletedAt *time.Time
+}
