+++ pkg/domain/entity/user.go
@@ -14,16 +14,3 @@ type User struct {
 }
 
 type UserSlice []*User
-
-type Profile struct {
-	Name      string
-	Thumbnail string
-	Bio       string
-	Gender    int
-	Phone     string
-	Place     string
-	Birth     string
-	CreatedAt *time.Time
-	UpdatedAt *time.Time
-	DeletedAt *time.Time
-}
